Fix conclusion chapter
The conclusion chapter doesn't have any headings in it which broke some of my hack logic.
This fixes:
- The chapter page now appears in the HTML TOC
- The chapter page has a page header like all the other pages
- The previous page now correctly links to the chapter page

This also removes a behaviour where fragment links weren't emitted for the first heading on a page. This didn't make so much sense because there can be content before the first heading on a page.